Wrap glob in quotes in NPM test-coverage command
The NPM test command had the glob for finding test files wrapped in quotes
however the NPM test-coverage command did not. The wrapping of the glob pattern
in quotes ensures that the glob is passed to the NPM command as is vs. being
evaulated by the NPM command.

That is we have two commands NPM test -> mocha <glob> and we want to make
sure that the glob pattern isn't evaluated to soon. That is because it could
resulted in the wrong files being found and passed to mocha.

The NPM test-coverage command seemed to be working but it was different then
the test command so this change set is mostly about making sure the two
commands are consistant.