:art:Remove -R flag from NPM test-coverage command
This change set removes the -R flag from the NPM test-coverage command.

The -R flag specifies the reporter used by mocha. It was set to tell Mocha to
use spec but that is the default reporter. Given that the NPM test command
does _not_ use this flag this change set removes it's use from the NPM
test-coverage command to remove inconsistancy between the two commands.

It is a net no-op operation but makes things more clear without having to
look up the default reporter