Remove the -u flag from NPM test-coverage command
The NPM test-coverage command was telling mocha to use the exports interface.
That is the test file syntax that mocha was expecting when parsing test files
was to be the exports style, tests are defineds as module exports.

The NPM test command did not use this flag thus was defaulting to the BDD
interface. Given that the NPM test command used BDD and that the tests were
written in the BDD interface this flag in the NPM test-coverage command
was incorrect.

This caused an issue when I tried to import a helper module which exported
a function as when using the NPM test-coverage command it was picked up as
a test file.

This change set removes the flag from the NPM test-coverage command.