:art:Move root after to own file
The readme stated that the tests will drop the test keyspace after the test
runs completed but I could not easly see what was dropping the keyspace. I
found that there was a root level after hook, which runs at the end of the
entire test suite, which dropped the keyspace but that after hook was defined
within one of the test files; I needed to read each test file to find it.

To make things a bit easier to navigate this change set moves the root level
after hook out of the test file and into it's own stand alown _setup_ module.

This setup module gets picked up by the glob pattern and it stands out as
the test setup code.

This change also increases the timeout from the default 2s to 10s for
the after hook as it seems to take a bit of time to drop the keyspace.