:art:Add root before to remove duplicate connects
The tests in the integration folder each connected to Cassandra however they
did this out side of a test block (describe) thus they all ran as the same
time when the tests were loaded, not when they were run. This meant that
we were telling cassie-odm to connect to cassandra multiple times in a row.

This also ment that we had a bunch of duplication of the connection config
as each test file defined it's won connection config object. This did cause
me a bit of issue as I do not have Cassandra installed locally which meant
I had to modify each test file in order to run the suite.

This change set refactors the integration tests to pull the same
configuration config object via the cassieconnect module and moved the
connecting/syncing from the test file's root to within the test block
so that if we skip the test we're not still connecting/syncing.

This change also adds a root before hook which will connect to
Cassandra once when the test suite starts negating the need for each
test file to connect to Cassandra.