Add Travis CI file
Travis CI (travis-ci.org) is a free hosted CI server which supports many
programing languages and platforms. It also has many services including
access to a Cassandra database.

To use Travis CI just define a .travis.yaml file which outlines how to setup
your testing environment and how to run tests then allow Travis to monitor
your GitHub repo.

This change set adds a Travis-CI file which will run the test suite against
multiple versions of Node.

Note: Travis will run npm install and npm test by default so this file is
pretty light. It just defines that we need access to a Cassandra database
and which version of node to run the tests against.

I picked each LTS version of Node from 0.10 as the package.json for this
project states that it can run on any version of Node from 0.10 and up.