:green_heart:Ensure raw query test will fail
Currently the raw query test will timeout on error, this is because we
do not tell mocha the test is done if the execute method return an
error. In that case we were just console logging which would result in
a test time out.

This change set updates the test to ensure we call the done callback with
the err object if the code under test returns a failure; this will result
in the test failing with a proper error message as to why the test failed
vs. just timing out.

It also removes the import of the assert module as we do not use it and
comments out the connect config line in favour of pulling it from the
cassieconnect module. This I think cleans the test up a bit as the only
changes between the test and the README is the commented out lines.