:green_heart:Refactor Type tests to assert on values
This change set refactors the Types tests to not just log exceptions
during save but to validate that the save and find operation saved
and returned the correct values as expected.

That is if we saved '42' then we should get '42' back.

It also updates the tests to fail the test if the save operation
failes vs. just logging it and marking the test passed

It also removes the long_type_tests.js file and moves the tests
into the types_test.js file under the Long test suite.