:bug:Fix bug in Long.toString()
This change set fixes a bug where we were passing an object into the long
constructor but it was expecting two seperate input parameters (low/high).

This change just pulls the low/high out of the input object when passing into
the long constructor. This keeps the API to the Long.toString() method
unchanged.