:arrow_up:Add chai and update schema tests
The chai module is an assertion library which has a lot more assertion APIs
then the build in Node Assertion module. Chai also has an assertion syntax
which allows for BDD test writting.

Notabily in regards to the Schema tests the Chai APIs do not perform auto
converstion of types so if the tests expects strings and the code returns
a number the native Assertion module would auto convert and pass the test
where as Chai will fail due to a type miss-match.

Chai also provides nicer APIs for checking types:
    expect(thing).to.be.a('function')

By using the chai library it makes it easier to express in the tests what
the expectations are for the code being tested.