:racehorse:Refactor Schema class to use the prototype
The Schema class each time it is instanitated would generate each method
anew. That is each time you newed up a Schema the constructor would instanitate
a new function for each method. This is not very efficent since the methods
on each Schema object are exactly the same. This results in a larger memory
foot print and a slower construction time.

I ran into this when profiling a fetch all route that returned 100's of
records, in that case it was the Model however, due to the re-creation
of each method for each row my route was orders of magantue slower then
when I mypassed the Model all together with a raw Query.

To improve the performance of the Schema model this change set refactors
the Schema class to use the Prototype.

In JavaScript we have prototypical classes, that is each object shares the
same prototype, all common _class_ methods/properties can be attached to
the prototype allowing each object to share the same instance of said
methods. These methods would act on the _this_ which would be each unique
object.

This means we only need one instance of the pre method for all instances
of the Schema class to share.

Note: I also changed the throws to throw an Error object vs. a string which
is just better Node practice.