:racehorse:Refactor Query class to use the prototype
Each time the Query class instantiates an instance of it self
the constructor function generates new instances of all the
methods of the class. This increases the memory foot print
and construction time.

To improve performance of the Query class this change set
refactors it to attach the methods, which will be the
same for each instance of the class, to the classes
prototype allowing each insatnace to share the same
methods vs. making new ones for each instance.