:bug:Return query object in findById
The findById function did not return the query object when no callback is
provided and did not throw if no callback is provided so it likely was
intended to return a query object like all the other find functions do.

This change set ensures that the findById function behaves like the other
find functions and returns the qeury object if no callback is passed in.