:bug: Rest dirty state when initializing an array
If the field is defined as an array we initialize it to an empty array so
consumers of the model can just push values into it; however, the custom
setter for that field property will set the dirty flag to true once we
set it to the empty array. This makes any empty model with an array field
'dirty' by default. To rectify that this change set re-sets the dirty
field after initializing the array field.