:bug: Initialize the map field
If the field is defined as a map consumers would expect to be able to attach
new properties (i.e. insert new entries into the map) however by default the
map field is left undefined. To provide a ready to use map field this change
set initializes map fields to an empty object.